Move certificate to documents endpoint
