+++ documents/urls.py
@@ -6,4 +6,5 @@ app_name = "documents"
 
 urlpatterns = [
     path("<uuid:pk>/", views.DocumentDetail.as_view(), name="document"),
+    path("certificate/", views.DownloadSigningCertificate.as_view(), name="certificate"),
 ]
